last changes (i hope)

+++ src/service/interfaces/BaseService.ts
@@ -1,4 +1,6 @@
 import { DeepPartial } from "typeorm";
+import { crewman } from "../../../test/mockEntities";
+import Crew from "../../model/Crew";
 import { BaseRepository, IBaseRepository } from "../../repository/interfaces/BaseRepository";
 
 export interface IBaseService<T> {
@@ -21,18 +23,22 @@ export abstract class BaseService<T>{
     }
 
     async getById(id: number): Promise<T> {
+        if (!id) throw new Error("id is required");
         return this.repository.getById(id);
     }
 
     async deleteById(id: number): Promise<T> {
+        if (!id) throw new Error("id is required");
         return this.repository.deleteById(id);
     }
 
-    async create(obj: DeepPartial<T>): Promise<T> {
+    async create(obj: DeepPartial<T> | T): Promise<T> {
         return this.repository.create(obj);
-    } 
+    }
 
     async update(id: number, obj: Partial<T>): Promise<T> {
+        if (!id) throw new Error("id is required");
+        if (!obj) throw new Error("obj is required");
         return this.repository.update(id, obj);
     }
 }
